Refactor common commands to shared crate
