+++ client/src/util.rs
@@ -59,49 +59,6 @@ pub fn init_logger(verbosity: u64) {
     log::set_logger(&LOGGER).unwrap();
 }
 
-pub fn human_duration(duration: Duration) -> String {
-    match duration.as_secs() {
-        n if n < 1 => "just now".cyan().to_string(),
-        n if n < 60 => format!("{} {} ago", n, "seconds".cyan()),
-        n if n < 60 * 60 => {
-            let mins = n / 60;
-            let secs = n % 60;
-            format!(
-                "{} {}, {} {} ago",
-                mins,
-                if mins == 1 { "minute" } else { "minutes" }.cyan(),
-                secs,
-                if secs == 1 { "second" } else { "seconds" }.cyan(),
-            )
-        },
-        n => {
-            let hours = n / (60 * 60);
-            let mins = (n / 60) % 60;
-            format!(
-                "{} {}, {} {} ago",
-                hours,
-                if hours == 1 { "hour" } else { "hours" }.cyan(),
-                mins,
-                if mins == 1 { "minute" } else { "minutes" }.cyan(),
-            )
-        },
-    }
-}
-
-pub fn human_size(bytes: u64) -> String {
-    const KB: u64 = 1024;
-    const MB: u64 = 1024 * KB;
-    const GB: u64 = 1024 * MB;
-    const TB: u64 = 1024 * GB;
-    match bytes {
-        n if n < 2 * KB => format!("{} {}", n, "B".cyan()),
-        n if n < 2 * MB => format!("{:.2} {}", n as f64 / KB as f64, "KiB".cyan()),
-        n if n < 2 * GB => format!("{:.2} {}", n as f64 / MB as f64, "MiB".cyan()),
-        n if n < 2 * TB => format!("{:.2} {}", n as f64 / GB as f64, "GiB".cyan()),
-        n => format!("{:.2} {}", n as f64 / TB as f64, "TiB".cyan()),
-    }
-}
-
 pub fn permissions_helptext(config_dir: &Path, data_dir: &Path, e: &io::Error) {
     if e.raw_os_error() == Some(1) {
         let current_exe = std::env::current_exe()
