+++ shared/src/types.rs
@@ -1,5 +1,4 @@
 use anyhow::{anyhow, Error};
-use clap::Args;
 use ipnet::IpNet;
 use lazy_static::lazy_static;
 use regex::Regex;
@@ -16,8 +15,7 @@ use std::{
 };
 use url::Host;
 use wireguard_control::{
-    AllowedIp, Backend, InterfaceName, InvalidInterfaceName, Key, PeerConfig, PeerConfigBuilder,
-    PeerInfo,
+    AllowedIp, InterfaceName, InvalidInterfaceName, Key, PeerConfig, PeerConfigBuilder, PeerInfo,
 };
 
 use crate::wg::PeerInfoExt;
@@ -281,198 +279,6 @@ impl<'a> CidrTree<'a> {
 #[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
 pub struct RedeemContents {
     pub public_key: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Args)]
-pub struct InstallOpts {
-    /// Set a specific interface name
-    #[clap(long, conflicts_with = "default-name")]
-    pub name: Option<String>,
-
-    /// Use the network name inside the invitation as the interface name
-    #[clap(long = "default-name")]
-    pub default_name: bool,
-
-    /// Delete the invitation after a successful install
-    #[clap(short, long)]
-    pub delete_invite: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Args)]
-pub struct AddPeerOpts {
-    /// Name of new peer
-    #[clap(long)]
-    pub name: Option<Hostname>,
-
-    /// Specify desired IP of new peer (within parent CIDR)
-    #[clap(long, conflicts_with = "auto-ip")]
-    pub ip: Option<IpAddr>,
-
-    /// Auto-assign the peer the first available IP within the CIDR
-    #[clap(long = "auto-ip")]
-    pub auto_ip: bool,
-
-    /// Name of CIDR to add new peer under
-    #[clap(long)]
-    pub cidr: Option<String>,
-
-    /// Make new peer an admin?
-    #[clap(long)]
-    pub admin: Option<bool>,
-
-    /// Bypass confirmation
-    #[clap(long)]
-    pub yes: bool,
-
-    /// Save the config to the given location
-    #[clap(long)]
-    pub save_config: Option<String>,
-
-    /// Invite expiration period (eg. '30d', '7w', '2h', '60m', '1000s')
-    #[clap(long)]
-    pub invite_expires: Option<Timestring>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Args)]
-pub struct RenamePeerOpts {
-    /// Name of peer to rename
-    #[clap(long)]
-    pub name: Option<Hostname>,
-
-    /// The new name of the peer
-    #[clap(long)]
-    pub new_name: Option<Hostname>,
-
-    /// Bypass confirmation
-    #[clap(long)]
-    pub yes: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Args)]
-pub struct AddCidrOpts {
-    /// The CIDR name (eg. 'engineers')
-    #[clap(long)]
-    pub name: Option<Hostname>,
-
-    /// The CIDR network (eg. '10.42.5.0/24')
-    #[clap(long)]
-    pub cidr: Option<IpNet>,
-
-    /// The CIDR parent name
-    #[clap(long)]
-    pub parent: Option<String>,
-
-    /// Bypass confirmation
-    #[clap(long)]
-    pub yes: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Args)]
-pub struct DeleteCidrOpts {
-    /// The CIDR name (eg. 'engineers')
-    #[clap(long)]
-    pub name: Option<String>,
-
-    /// Bypass confirmation
-    #[clap(long)]
-    pub yes: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Args)]
-pub struct AddDeleteAssociationOpts {
-    /// The first cidr to associate
-    pub cidr1: Option<String>,
-
-    /// The second cidr to associate
-    pub cidr2: Option<String>,
-
-    /// Bypass confirmation
-    #[clap(long)]
-    pub yes: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Args)]
-pub struct ListenPortOpts {
-    /// The listen port you'd like to set for the interface
-    #[clap(short, long)]
-    pub listen_port: Option<u16>,
-
-    /// Unset the local listen port to use a randomized port
-    #[clap(short, long, conflicts_with = "listen-port")]
-    pub unset: bool,
-
-    /// Bypass confirmation
-    #[clap(long)]
-    pub yes: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Args)]
-pub struct OverrideEndpointOpts {
-    /// The listen port you'd like to set for the interface
-    #[clap(short, long)]
-    pub endpoint: Option<Endpoint>,
-
-    /// Unset an existing override to use the automatic endpoint discovery
-    #[clap(short, long, conflicts_with = "endpoint")]
-    pub unset: bool,
-
-    /// Bypass confirmation
-    #[clap(long)]
-    pub yes: bool,
-}
-
-#[derive(Debug, Clone, Args)]
-pub struct NatOpts {
-    #[clap(long)]
-    /// Don't attempt NAT traversal. Note that this still will report candidates
-    /// unless you also specify to exclude all NAT candidates.
-    pub no_nat_traversal: bool,
-
-    #[clap(long)]
-    /// Exclude one or more CIDRs from NAT candidate reporting.
-    /// ex. --exclude-nat-candidates '0.0.0.0/0' would report no candidates.
-    pub exclude_nat_candidates: Vec<IpNet>,
-
-    #[clap(long, conflicts_with = "exclude-nat-candidates")]
-    /// Don't report any candidates to coordinating server.
-    /// Shorthand for --exclude-nat-candidates '0.0.0.0/0'.
-    pub no_nat_candidates: bool,
-}
-
-impl NatOpts {
-    pub fn all_disabled() -> Self {
-        Self {
-            no_nat_traversal: true,
-            exclude_nat_candidates: vec![],
-            no_nat_candidates: true,
-        }
-    }
-
-    /// Check if an IP is allowed to be reported as a candidate.
-    pub fn is_excluded(&self, ip: IpAddr) -> bool {
-        self.no_nat_candidates
-            || self
-                .exclude_nat_candidates
-                .iter()
-                .any(|network| network.contains(&ip))
-    }
-}
-
-#[derive(Debug, Clone, Copy, Args)]
-pub struct NetworkOpts {
-    #[clap(long)]
-    /// Whether the routing should be done by innernet or is done by an
-    /// external tool like e.g. babeld.
-    pub no_routing: bool,
-
-    #[clap(long, default_value_t, possible_values = Backend::variants())]
-    /// Specify a WireGuard backend to use.
-    /// If not set, innernet will auto-select based on availability.
-    pub backend: Backend,
-
-    #[clap(long)]
-    /// Specify the desired MTU for your interface (default: 1280).
-    pub mtu: Option<u32>,
 }
 
 #[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
